re-organized package hierarchy
