--- src/bomberman/animations/Animations.java
@@ -1,7 +1,0 @@
-package bomberman.animations;
-
-import bomberman.animations.sprites.Sprite;
-
-public interface Animations {
-	public Sprite getSprite();
-}
