+++ src/bomberman/entity/bomb/Explosion.java
@@ -3,11 +3,11 @@ package bomberman.entity.bomb;
 import java.util.Date;
 
 import bomberman.Sandbox;
-import bomberman.animations.SingleSpriteAnimations;
-import bomberman.animations.sprites.SpriteSpecification;
 import bomberman.constants.EntityDimensions;
 import bomberman.entity.Entity;
-import bomberman.entity.boundedbox.CollidableType;
+import bomberman.entity.configurations.animations.SingleSpriteAnimations;
+import bomberman.entity.configurations.animations.sprites.SpriteSpecification;
+import bomberman.entity.configurations.boundedbox.CollidableType;
 
 public class Explosion extends Entity {
