+++ src/bomberman/entity/configurations/animations/Animations.java
@@ -1,0 +1,13 @@
+package bomberman.entity.configurations.animations;
+
+import bomberman.entity.configurations.animations.sprites.Sprite;
+
+
+/**
+ * Used by each entity to hold on to all sprites related to it
+ * @author tialim
+ */
+
+public interface Animations {
+	public Sprite getSprite();
+}
