+++ src/bomberman/entity/configurations/animations/SingleSpriteAnimations.java
@@ -1,11 +1,12 @@
-package bomberman.animations;
+package bomberman.entity.configurations.animations;
 
-import bomberman.animations.sprites.Sprite;
-import bomberman.animations.sprites.SpriteSpecification;
 import bomberman.entity.Entity;
+import bomberman.entity.configurations.animations.sprites.Sprite;
+import bomberman.entity.configurations.animations.sprites.SpriteSpecification;
 
-/* 
+/**
  * For entities that have only one sprite
+ * @author tialim
  */
 //TODO better descriptor
