+++ src/bomberman/entity/killableentity/player/Player.java
@@ -1,14 +1,13 @@
 package bomberman.entity.killableentity.player;
 
 import bomberman.Sandbox;
-import bomberman.animations.PlayerAnimations;
 import bomberman.constants.EntityDimensions;
 import bomberman.constants.GlobalConstants;
 import bomberman.entity.Entity;
-import bomberman.entity.bomb.BlackBomb;
-import bomberman.entity.boundedbox.CollidableType;
-import bomberman.entity.boundedbox.RectBoundedBox;
-import bomberman.entity.movements.MovingEntity;
+import bomberman.entity.configurations.animations.PlayerAnimations;
+import bomberman.entity.configurations.boundedbox.CollidableType;
+import bomberman.entity.configurations.boundedbox.RectBoundedBox;
+import bomberman.entity.configurations.movements.MovingEntity;
 import bomberman.gamecontroller.InputManager;
 
 public class Player extends MovingEntity {
