+++ src/resources/img/sprites_without_border.png
